# api/src/main/java/com/unicar/docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: UniCar API
  description: >
    API do sistema UniCar — plataforma de caronas universitárias.
    Cobre autenticação, usuários, veículos, caronas, reservas, avaliações,
    comunicação, busca e trajetos recorrentes.
  version: 1.0.0
servers:
  - url: https://api.unicar.example.com
    description: Servidor de produção
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Usuario:
      type: object
      properties:
        id:
          type: integer
          example: 1
        matricula:
          type: string
          example: "123456789"
        nome:
          type: string
          example: Jennifer Medeiros
        email:
          type: string
          format: email
          example: jennifer@ccc.ufcg.edu.br
        cpf:
          type: string
          example: "12345678900"
        curso:
          type: string
          example: Ciência da Computação
        genero:
          type: string
          enum: [MASCULINO, FEMININO, NAO_INFORMADO]
          example: FEMININO
        receberEmail:
          type: boolean
          example: true
        dataCriacao:
          type: string
          format: date-time
          example: "2026-06-22T10:00:00"
        dataAtualizacao:
          type: string
          format: date-time
          example: "2026-06-22T10:00:00"
    UsuarioPublico:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        matricula:
          type: string
          example: "2023001234"
        nome:
          type: string
          example: João da Silva
        email:
          type: string
          format: email
          example: joao.silva@academico.ufcg.edu.br
        curso:
          type: string
          example: Ciência da Computação
...
    AtualizarPerfilRequest:
      type: object
      properties:
        genero:
          type: string
          enum: [MASCULINO, FEMININO, NAO_INFORMADO]
          example: FEMININO
        receberEmail:
          type: boolean
          example: false

    # ─── Bloqueio ─────────────────────────────────────────────────────────────
    Bloqueio:
      type: object
      properties:
        id:
          type: integer
          example: 10
        usuarioId:
          type: integer
          example: 1
        usuarioBloqueadoId:
          type: integer
          example: 5
        dataBloqueio:
          type: string
          format: date-time
          example: "2026-06-22T10:00:00"
    UsuarioBloqueado:
      type: object
      properties:
        id:
          type: integer
          example: 5
        nome:
          type: string
          example: João Silva
        curso:
          type: string
          example: Engenharia Elétrica
        dataBloqueio:
          type: string
          format: date-time
          example: "2026-06-22T10:00:00"

    # ─── Avaliação ────────────────────────────────────────────────────────────
...
    AvaliacaoRecebida:
      type: object
      properties:
        id:
          type: integer
          example: 100
        nota:
          type: integer
          example: 5
        comentario:
          type: string
          example: Motorista pontual e educado.
        dataAvaliacao:
          type: string
          format: date-time
          example: "2026-06-25T18:30:00"
        avaliador:
          type: object
          properties:
            id:
              type: integer
              example: 10
            nome:
              type: string
              example: João Silva
        carona:
          type: object
          properties:
            id:
              type: integer
              example: 25
...
    Reputacao:
      type: object
      properties:
        usuarioId:
          type: integer
          example: 5
        media:
          type: number
          format: float
          example: 4.8
        quantidadeAvaliacoes:
          type: integer
          example: 35
...
    PerfilPublico:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Maria Oliveira
        curso:
          type: string
          example: Ciência da Computação
        genero:
          type: string
          example: FEMININO
        reputacao:
          type: number
          format: float
          example: 4.8
        quantidadeAvaliacoes:
          type: integer
          example: 15
...
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
          example: Mensagem de erro

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /usuarios/me:
    get:
      tags: [Usuários]
      summary: Consultar perfil do usuário autenticado
      description: >
        Retorna os dados do usuário autenticado a partir do token JWT enviado
        no header Authorization (Bearer Token) (RN-AUTH-04).
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dados do usuário autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '401':
          description: Usuário não autenticado ou token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /usuarios/{matricula}:
    get:
      tags: [Usuários]
      summary: Consulta dados públicos de um usuário
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: matricula
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dados públicos do usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioPublico'
        '404':
          description: Usuário não encontrado

    patch:
      tags: [Usuários]
      summary: Atualizar perfil do usuário autenticado
      description: >
        Permite ao usuário autenticado atualizar apenas os campos
        genero e receberEmail do próprio perfil
        (RN-AUTH-05 a RN-AUTH-07).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizarPerfilRequest'
      responses:
        '200':
          description: Perfil atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '401':
          description: Usuário não autenticado ou token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'

    delete:
      tags: [Usuários]
      summary: Desativar cadastro do usuário autenticado
      description: >
        Desativa o cadastro do próprio usuário no UniCar sem remover
        histórico e sem afetar a conta institucional no Eureca
        (RN-AUTH-08 a RN-AUTH-14).
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Usuário desativado com sucesso
        '401':
          description: Usuário não autenticado ou token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /usuarios/me/avaliacoes:
    get:
      tags: [Avaliações]
      summary: Consultar avaliações recebidas pelo usuário autenticado
      description: >
        Retorna avaliações recebidas com média e quantidade calculadas
        automaticamente (RN-AVA-08, RN-AVA-09). Também descrito em US4
        como resumo (RN-AVA-01 a RN-AVA-03).
      responses:
        '200':
          description: Lista de avaliações recebidas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AvaliacaoRecebida'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /usuarios/bloqueados:
    get:
      tags: [Segurança]
      summary: Listar usuários bloqueados
      description: >
        Retorna todos os usuários bloqueados pelo usuário autenticado
        (RN-BLOQ-05).
      responses:
        '200':
          description: Lista de usuários bloqueados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioBloqueado'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /usuarios/{id}/bloquear:
    post:
      tags: [Segurança]
      summary: Bloquear um usuário
      description: >
        Bloqueia o usuário identificado por `id`. Usuários bloqueados não
        interagem entre si, não aparecem em buscas e não podem trocar mensagens
        (RN-BLOQ-01 a RN-BLOQ-03).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 5
      responses:
        '201':
          description: Bloqueio registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bloqueio'
        '400':
          description: Usuário já bloqueado ou tentativa de bloquear a si mesmo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'

    delete:
      tags: [Segurança]
      summary: Remover bloqueio de um usuário
      description: Remove o bloqueio previamente criado (RN-BLOQ-04).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 5
      responses:
        '204':
          description: Bloqueio removido
        '404':
          description: Bloqueio não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /usuarios/{id}/perfil-publico:
    get:
      tags: [Busca]
      summary: Consultar perfil público de um usuário
      description: >
        Retorna apenas informações públicas do usuário. Respeita bloqueios entre
        usuários (RN-BUS-09 a RN-BUS-11).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Perfil público
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerfilPublico'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /usuarios/{id}/reputacao:
    get:
      tags: [Avaliações]
      summary: Consultar reputação pública de um usuário
      description: >
        Retorna a média das avaliações recebidas e o total de avaliações
        (RN-AVA-10 a RN-AVA-12).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 5
      responses:
        '200':
          description: Reputação do usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reputacao'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'

  # ── Notificações ──────────────────────────────────────────────────────────
...